fix: 리액티브 redis에서 톰캣 기반 redis 설정 기반 리팩토링
