+++ src/main/java/com/example/chatserverauth/global/config/ReactiveRedisConfig.java
@@ -1,69 +1,69 @@
-package com.example.chatserverauth.global.config;
-
-import com.example.chatserverauth.domain.dto.UserInfoDTO;
-import io.lettuce.core.ClientOptions;
-import io.lettuce.core.SocketOptions;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
-import org.springframework.boot.autoconfigure.data.redis.RedisReactiveAutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceClientConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.ReactiveRedisTemplate;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-import java.time.Duration;
-
-@Configuration
-@EnableAutoConfiguration(exclude={RedisAutoConfiguration.class, RedisReactiveAutoConfiguration.class}) // 자동 생성 설정과 커스텀 설정 충돌 방지
-public class ReactiveRedisConfig {
-
-    @Value("${spring.data.redis.host}")
-    private String host;
-    @Value("${spring.data.redis.port}")
-    private int port;
-    @Value("${spring.data.redis.password}")
-    private String password;
-
-    @Bean
-    // Lettuce는 ReactiveRedisConnectionFactory의 구현체 중 일부
-    public ReactiveRedisConnectionFactory connectionFactory() {
-        RedisStandaloneConfiguration redisConfiguration = new RedisStandaloneConfiguration();
-        redisConfiguration.setHostName(host);
-        redisConfiguration.setPort(port);
-        redisConfiguration.setPassword(password);
-        redisConfiguration.setDatabase(0);
-
-        final SocketOptions socketoptions = SocketOptions.builder().connectTimeout(Duration.ofSeconds(10)).build();
-        final ClientOptions clientoptions = ClientOptions.builder().socketOptions(socketoptions).build();
-
-        LettuceClientConfiguration lettuceClientConfiguration = LettuceClientConfiguration.builder()
-                .clientOptions(clientoptions)
-                .commandTimeout(Duration.ofMinutes(1))
-                .shutdownTimeout(Duration.ZERO)
-                .build();
-
-        return new LettuceConnectionFactory(redisConfiguration, lettuceClientConfiguration);
-    }
-
-    @Bean(name = "userInfoTemplate")
-    public ReactiveRedisTemplate<String, UserInfoDTO> reactiveRedisTemplate(ReactiveRedisConnectionFactory connectionFactory) {
-        Jackson2JsonRedisSerializer<UserInfoDTO> serializer = new Jackson2JsonRedisSerializer<>(UserInfoDTO.class);
-        RedisSerializationContext.RedisSerializationContextBuilder<String, UserInfoDTO> builder
-                = RedisSerializationContext.newSerializationContext(new StringRedisSerializer());
-
-        RedisSerializationContext<String, UserInfoDTO> context = builder.value(serializer)
-                .hashValue(serializer)
-                .hashKey(serializer)
-                .build();
-
-        return new ReactiveRedisTemplate<>(connectionFactory, context);
-    }
-}
+//package com.example.chatserverauth.global.config;
+//
+//import com.example.chatserverauth.domain.dto.UserInfoDTO;
+//import io.lettuce.core.ClientOptions;
+//import io.lettuce.core.SocketOptions;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+//import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
+//import org.springframework.boot.autoconfigure.data.redis.RedisReactiveAutoConfiguration;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.context.annotation.Primary;
+//import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
+//import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+//import org.springframework.data.redis.connection.lettuce.LettuceClientConfiguration;
+//import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+//import org.springframework.data.redis.core.ReactiveRedisTemplate;
+//import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
+//import org.springframework.data.redis.serializer.RedisSerializationContext;
+//import org.springframework.data.redis.serializer.StringRedisSerializer;
+//
+//import java.time.Duration;
+//
+//@Configuration
+//@EnableAutoConfiguration(exclude={RedisAutoConfiguration.class, RedisReactiveAutoConfiguration.class}) // 자동 생성 설정과 커스텀 설정 충돌 방지
+//public class ReactiveRedisConfig {
+//
+//    @Value("${spring.data.redis.host}")
+//    private String host;
+//    @Value("${spring.data.redis.port}")
+//    private int port;
+//    @Value("${spring.data.redis.password}")
+//    private String password;
+//
+//    @Bean
+//    // Lettuce는 ReactiveRedisConnectionFactory의 구현체 중 일부
+//    public ReactiveRedisConnectionFactory connectionFactory() {
+//        RedisStandaloneConfiguration redisConfiguration = new RedisStandaloneConfiguration();
+//        redisConfiguration.setHostName(host);
+//        redisConfiguration.setPort(port);
+//        redisConfiguration.setPassword(password);
+//        redisConfiguration.setDatabase(0);
+//
+//        final SocketOptions socketoptions = SocketOptions.builder().connectTimeout(Duration.ofSeconds(10)).build();
+//        final ClientOptions clientoptions = ClientOptions.builder().socketOptions(socketoptions).build();
+//
+//        LettuceClientConfiguration lettuceClientConfiguration = LettuceClientConfiguration.builder()
+//                .clientOptions(clientoptions)
+//                .commandTimeout(Duration.ofMinutes(1))
+//                .shutdownTimeout(Duration.ZERO)
+//                .build();
+//
+//        return new LettuceConnectionFactory(redisConfiguration, lettuceClientConfiguration);
+//    }
+//
+//    @Bean(name = "userInfoTemplate")
+//    public ReactiveRedisTemplate<String, UserInfoDTO> reactiveRedisTemplate(ReactiveRedisConnectionFactory connectionFactory) {
+//        Jackson2JsonRedisSerializer<UserInfoDTO> serializer = new Jackson2JsonRedisSerializer<>(UserInfoDTO.class);
+//        RedisSerializationContext.RedisSerializationContextBuilder<String, UserInfoDTO> builder
+//                = RedisSerializationContext.newSerializationContext(new StringRedisSerializer());
+//
+//        RedisSerializationContext<String, UserInfoDTO> context = builder.value(serializer)
+//                .hashValue(serializer)
+//                .hashKey(serializer)
+//                .build();
+//
+//        return new ReactiveRedisTemplate<>(connectionFactory, context);
+//    }
+//}
